flask-api-service/requirements.txt
Flask==2.2.3
Flask-CORS==3.0.10
psycopg2==2.9.3
Werkzeug==2.2.2
requests==2.28.1
python-dotenv==0.20.0